Refactored mmDirSnd Moved mmDirSnd to mm2_sound.h
