+++ src/mm2.h
@@ -5,6 +5,7 @@
     Class definitions
 */
 #include "mm2_ai.h"
+#include "mm2_audio.h"
 #include "mm2_base.h"
 #include "mm2_city.h"
 #include "mm2_data.h"
