+++ src/mm2_audio.cpp
@@ -1,0 +1,2 @@
+#include "mm2_audio.h"
+using namespace MM2;
